+++ Ion/lib/ms5607.h
@@ -29,5 +29,5 @@ char readUInt_16(uint8_t address, uint16_t *value);
 char readBytes(unsigned char *values, int length);
 char startConversion(char CMD);
 char startMeasurment(void);
-char getDigitalValue(uint64_t *value);
+char getDigitalValue(uint32_t *value);
